File structure cosmetic change.
