+++ src/routes/demos/highlighterDemo.ts
